REF move virtual packages to their own module
