+++ conda_forge_feedstock_check_solvable/check_solvable.py
@@ -6,10 +6,7 @@ import conda_build.api
 import psutil
 from ruamel.yaml import YAML
 
-from conda_forge_feedstock_check_solvable.mamba_solver import (
-    _mamba_factory,
-    virtual_package_repodata,
-)
+from conda_forge_feedstock_check_solvable.mamba_solver import _mamba_factory
 from conda_forge_feedstock_check_solvable.utils import (
     MAX_GLIBC_MINOR,
     apply_pins,
@@ -19,6 +16,9 @@ from conda_forge_feedstock_check_solvable.utils import (
     print_warning,
     remove_reqs_by_name,
     suppress_output,
+)
+from conda_forge_feedstock_check_solvable.virtual_packages import (
+    virtual_package_repodata,
 )
